FPM enroll function.

+++ lock/fpm.h
@@ -33,6 +33,6 @@ uint16_t fpm_get_count(void);
 
 uint8_t fpm_enroll(void);
 
-uint8_t fpm_match(void);
+uint16_t fpm_match(void);
 
 #endif /* FPM_R50_H */
